사이드바오류디비오류수정사이드트리맵

+++ Ttree/app/controllers/apis_controller.rb
@@ -96,9 +96,14 @@ class ApisController < ApplicationController
 
 	def tree
 		@work=Work.find(params[:id])
+		#@branches=Branch.all
+		#mappings = {"name" => "title"}
+
 		@branches=@work.branches.arrange_serializable
-		#render :json =>  Branch.json_tree(@branches)
-		render :json => @branches
+		#@branches=@work.branches
+		#render :json =>  Branch.json_search(@branches)
+		
+		render :json => @branches 
 	end
 
 	def branchChilds
@@ -145,9 +150,12 @@ class ApisController < ApplicationController
 	def getPages
 		@json= JSON.parse(request.raw_post)
 		@user= User.find_by_email(@json["user_email"])
+		@time=Time.now
 		@json["pages"].each do |page|
-			Unclassifiedpage.create(:user_id=>@user.id, :title=>page["title"], :url=>page["url"])
+			@thisPage=Unclassifiedpage.create(:user_id=>@user.id, :title=>page["title"], :url=>page["url"], :timenum=>@time)
+			#@thisPage.update_attributes(:timenum => @time)
 		end
+		#debugger
 	end
 
 	def getMember
@@ -170,12 +178,19 @@ class ApisController < ApplicationController
 	def branchName
 		@name=branch_params
 		@branch=Branch.create(branch_params)
-		@pages=User.current.unclassifiedpage_ids
-		@pages.each do |page|
-			@page=Unclassifiedpage.find(page)
-			@branch.pages.create(title: @page.title, url:@page.url)
+		@pages=User.find(params[:user_id]).unclassifiedpage_ids
+		Unclassifiedpage.transaction do
+			@pages.each do |page|
+				@page=Unclassifiedpage.find(page)
+				@branch.transaction do
+					@branch.pages.create(title: @page.title, url:@page.url)
+				end
+			end  
 		end
-		Unclassifiedpage.delete_all
+		
+		Unclassifiedpage.transaction do
+			Unclassifiedpage.where(:user_id=>params[:user_id]).delete_all
+		end
 		#debugger
 	end
 
@@ -212,6 +227,11 @@ class ApisController < ApplicationController
 	def deletePages
 		Page.find(params[:id]).destroy
 	end
+
+	def treeSideBar
+		render :json => '[{ "id" : "ajson1", "parent" : "#", "text" : "Simple root node" },{ "id" : "ajson2", "parent" : "#", "text" : "Root node 2" },{ "id" : "ajson3", "parent" : "ajson2", "text" : "Child 1" },{ "id" : "ajson4", "parent" : "ajson2", "text" : "Child 2" }]'
+	end
+
 
 	private
 	def work_params
